requirements.txt
discord.py
mysql-connector-python
discodo
youtube_dl